carrinho funcionando, além disso, agora tem um autenticador de sessão, pequenas mudanças feitas por todo o código, nada de muito relevante... Apenas o usuário e a senha do Banco de Dados postgres teve a senha e o usuário alterados ambos para 'postgres'

+++ pw/src/main/java/com/trabalho/pw/Produto.java
@@ -2,13 +2,13 @@ package com.trabalho.pw;
 
 public class Produto {
 
-    private int id;
+    private Long id;
     private String nome;
     private String descricao;
     private double preco;
     private int qntd;
 
-    public Produto(int id, String nome, String descricao, double preco, int qntd) {
+    public Produto(Long id, String nome, String descricao, double preco, int qntd) {
         this.id = id;
         this.nome = nome;
         this.descricao = descricao;
@@ -16,7 +16,7 @@ public class Produto {
         this.qntd = qntd;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
